Fix clippy lints across workspace
Workspace-wide cleanup so cargo clippy --workspace --all-targets
-D warnings passes (commit gate required by plan + AGENTS.md
section 2). 40 files touched.

Categories:
- missing_docs: ~150 items (M8 chat + M9 payment + M14 User aggregate
  + common/config fields + infra ports + worker handlers)
- unused imports: 5 (mssql_payment.rs Decimal + read_i32/str/uuid,
  redis_chat.rs broadcast)
- unused field/method: cache/layer.rs (l1_max_ttl, l1_ttl - dead code)
- complex type: cache/layer.rs InflightMap (extracted type alias)
- manual RangeInclusive: cache/layer.rs (now (1..=3).contains(&runs))
- field_reassign_with_default: common/config.rs (DatabaseTopologySettings),
  infra/cache/redis.rs (live tests)
- derivable_impls: common/config.rs (Settings, DatabaseTopologySettings)
- lifetime elision: infra/db/mssql.rs read_str (dropped explicit 'a)
- needless_borrow: infra/db/mssql_user.rs (2x SpError::from_code)
- useless conversion: api/handlers/chat.rs + ws.rs (room_id.into()
  where RoomId is a Uuid type alias)
- unused mut: infra/pubsub/redis_chat.rs (get_async_connection)
- question_mark: api/handlers/order.rs + payment.rs (assert_role?)
- deprecated API: infra/pubsub/redis_chat.rs (intentionally pinned
  for pubsub stream; see #[allow(deprecated)] + comment)
- result_large_err: api/middleware/auth.rs assert_role (intentionally
  kept inline; boxing would break ? chain - see ponytail comment)

No logic changes. Pure docs + trivial mechanical fixes.

Unblocks T-01 commit (rustfmt.toml).

+++ crates/worker/src/idempotency.rs
@@ -28,6 +28,7 @@ pub struct IdempotencyKey {
 }
 
 impl IdempotencyKey {
+    /// Build a new key from the consumer name + upstream message id.
     pub fn new(consumer: impl Into<String>, message_id: impl Into<String>) -> Self {
         Self {
             consumer: consumer.into(),
@@ -42,9 +43,10 @@ impl IdempotencyKey {
     }
 }
 
+/// Errors raised by the idempotency cache backend.
 #[derive(Debug, Error)]
 pub enum IdempotencyError {
-    /// Cache backend failure.
+    /// Cache backend failure (Redis down, etc.).
     #[error("idempotency backend error: {0}")]
     Backend(String),
 }
@@ -66,6 +68,8 @@ pub struct InMemoryIdempotency {
 }
 
 impl InMemoryIdempotency {
+    /// Construct with a soft cap on cache size (`max_entries`); eviction
+    /// is naive (drops half) — see M4 note about LRU.
     pub fn new(max_entries: usize) -> Self {
         Self {
             seen: Arc::new(Mutex::new(std::collections::HashSet::new())),
@@ -100,6 +104,7 @@ pub struct RedisIdempotency {
 }
 
 impl RedisIdempotency {
+    /// Wrap a deadpool-redis pool (production multi-instance cache).
     pub fn new(pool: deadpool_redis::Pool) -> Self {
         Self { pool }
     }
@@ -114,7 +119,7 @@ impl Idempotency for RedisIdempotency {
             .await
             .map_err(|e| IdempotencyError::Backend(e.to_string()))?;
         let cache_key = key.cache_key();
-        let ttl_secs = ttl.as_secs().max(1) as u64;
+        let ttl_secs = ttl.as_secs().max(1);
         let result: Option<String> = redis::cmd("SET")
             .arg(&cache_key)
             .arg("1")
